requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
supabase==2.10.0
googlemaps==4.10.0
ortools
python-jose[cryptography]==3.3.0
python-multipart==0.0.12
python-dotenv==1.0.1
pytest==8.3.3

Pillow
numpy
google-generativeai